backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-socketio==5.10.0
python-engineio==4.8.0
requests==2.31.0
httpx==0.25.2
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
apscheduler==3.10.4
pydantic==2.5.2
pydantic-settings==2.1.0
aiofiles==23.2.1
python-multipart==0.0.6
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
celery==5.3.4
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1